backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.7
anthropic==0.7.7
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
httpx==0.25.2
python-socketio==5.10.0
websockets==12.0
aiofiles==23.2.1
Pillow==10.1.0